requirements.txt
# === Core ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-socketio==5.11.0
pydantic==2.10.3
pydantic-settings==2.7.0

# === Browser Automation ===
playwright==1.49.1

# === AI / LLM ===
openai==1.58.1
anthropic==0.42.0
tiktoken==0.8.0

# === Image Processing ===
Pillow==11.1.0
pixelmatch==0.3.0

# === Performance ===
aiohttp==3.11.11
httpx==0.28.1

# === Utilities ===
python-dotenv==1.0.1
rich==13.9.4
jinja2==3.1.5
python-multipart==0.0.20
orjson==3.10.13
pyyaml==6.0.2

# === Testing ===
pytest==8.3.4
pytest-asyncio==0.25.0

# === Export ===
weasyprint==63.1

# === MCP Server (for Copilot/Claude integration) ===
mcp[cli]>=1.0.0
